backend/pyproject.toml
[project]
name = "kubestats"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]>=0.115.0,<1.0.0",
    "python-multipart>=0.0.10,<1.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "tenacity>=9.0.0,<10.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "alembic>=1.15.0,<2.0.0",
    "httpx>=0.28.0,<1.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "sqlmodel>=0.0.21,<1.0.0",
    "bcrypt>=4.2.0,<5.0.0",
    "pydantic-settings>=2.9.0,<3.0.0",
    "pyjwt>=2.10.0,<3.0.0",
    "celery[redis]>=5.5.0,<6.0.0",
    "redis>=4.5.2,<4.5.5",
    "async-timeout>=4.0.2",
    "psutil>=7.0.0,<7.1.0",
    "gunicorn>=23.0.0,<24.0.0",
    "GitPython>=3.1.40,<4.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0,<9.0.0",
    "mypy>=1.13.0,<2.0.0",
    "ruff>=0.8.0,<1.0.0",
    "pre-commit>=4.0.0,<5.0.0",
    "types-passlib>=1.7.7.20241201,<2.0.0",
    "coverage>=7.6.0,<8.0.0",
    "pytest-env>=1.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic", "data"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic", "data"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["ARG001"]
"**/tests/**/*.py" = ["ARG001"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.pytest.ini_options]
env = [
    "POSTGRES_DB=test_db",
]
